changed files names & grouped into packages
