+++ UsersApp/Extensions/UsersImageExtensions.swift
@@ -1,0 +1,31 @@
+//
+//  ImageExtensions.swift
+//  UsersApp
+//
+//  Created by Dar Dar on 09.11.2023.
+//
+
+import UIKit
+
+
+extension UIImageView {
+    func downloaded(from url: URL, contentMode mode: ContentMode = .scaleToFill) {
+        contentMode = mode
+        URLSession.shared.dataTask(with: url) { data, response, error in
+            guard
+                let httpResponse = response as? HTTPURLResponse, httpResponse.statusCode == 200,
+                let mimeType = response?.mimeType, mimeType.hasPrefix("image"),
+                let data = data, error == nil,
+                let image = UIImage(data: data)
+            else{return}
+            DispatchQueue.main.async() { [weak self] in
+                self?.image = image
+            }
+        }.resume()
+    }
+    
+    func downloaded(from link: String, contentMode mode: ContentMode = .scaleToFill) {
+        guard let url = URL(string: link) else {return}
+        downloaded(from: url, contentMode: mode)
+    }
+}
